refactor(internal/database): seralize revisionId and entityKey
